Update Player.java
Checkbalance now returns float

+++ eclipse-workspace/devops_black_jack/src/devops_black_jack/Player.java
@@ -11,7 +11,7 @@ public class Player {
 	
 	
 	
-	public int checkBalance() {
+	public float checkBalance() {
 		
 		return balance;
 	}
